Add access to pending receipts.

+++ eth/query/block.go
@@ -19,11 +19,11 @@ var (
 	searchBlockSize = uint64(100)
 )
 
-func GetBlockByNumber(state loomchain.ReadOnlyState, height uint64, full bool, readReceipts loomchain.ReadReceiptHandler) ([]byte, error) {
+func GetBlockByNumber(state loomchain.ReadOnlyState, height int64, full bool, readReceipts loomchain.ReadReceiptHandler) ([]byte, error) {
 	params := map[string]interface{}{}
 	params["heightPtr"] = &height
 	var blockresult *ctypes.ResultBlock
-	iHeight := int64(height)
+	iHeight := height
 	blockresult, err := core.Block(&iHeight)
 	if err != nil {
 		return nil, err
@@ -34,16 +34,16 @@ func GetBlockByNumber(state loomchain.ReadOnlyState, height uint64, full bool, r
 
 		Timestamp: int64(blockresult.Block.Header.Time.Unix()),
 	}
-	if uint64(state.Block().Height) == height {
+	if state.Block().Height == height {
 		blockinfo.Number = 0
 	} else {
-		blockinfo.Number = int64(height)
+		blockinfo.Number = height
 	}
 
-	bloomFilter := common.GetBloomFilter(state, height)
+	bloomFilter := common.GetBloomFilter(state, uint64(height))
 	blockinfo.LogsBloom = bloomFilter
 
-	txHashList, err := common.GetTxHashList(state, height)
+	txHashList, err := common.GetTxHashList(state, uint64(height))
 	if err != nil {
 		return nil, errors.Wrap(err, "getting tx hash")
 	}
@@ -111,7 +111,7 @@ func GetBlockByHash(state loomchain.ReadOnlyState, hash []byte, full bool, readR
 		}
 		for i := int(len(info.BlockMetas) - 1); i >= 0; i-- {
 			if 0 == bytes.Compare(hash, info.BlockMetas[i].BlockID.Hash) {
-				return GetBlockByNumber(state, uint64(int(end)+i), full, readReceipts)
+				return GetBlockByNumber(state, int64(int(end)+i), full, readReceipts)
 			}
 		}
